Refactor: Introduced transition on header minification again.
The only confusing part was that we changed body top-margin all the time. I have isolated
that part of ajustHeaderHeight into ajustBodyMarginTop, and have decoupled it from scrollTop.
Also now calculating full header height from .navbar-header, not kb-navbar-container, so it
always gets the height of the content, not the (shifting) height of the container.